Using double

+++ targets/simple_switch/simple_switch.cpp
@@ -115,15 +115,15 @@ struct cmp_str {
    }
 };
 
-std::map<char*, std::tuple<time_t, uint32_t, double>, cmp_str> maxflow_map;
+std::map<char*, std::tuple<time_t, double, double>, cmp_str> maxflow_map;
 
 void calculate_drop_rate(char* route_id, char* packet_size, char* maxflow_handle) {
   if (!maxflow_map.count(route_id)) {
     std::cout << "No maxflow_map match " << std::endl;
-    maxflow_map[route_id] = std::make_tuple(time(0), 0, 0.0);
+    maxflow_map[route_id] = std::make_tuple(time(0), 0.0, 0.0);
   } else {
     std::cout << "Found maxflow_map match " << std::endl;
-    uint32_t& total_length = std::get<1>(maxflow_map[route_id]);
+    double& total_length = std::get<1>(maxflow_map[route_id]);
     total_length += ((packet_size[0] << 8) | packet_size[1]);
     time_t state_time = std::get<0>(maxflow_map[route_id]);
     time_t current_time = time(0);
@@ -137,7 +137,7 @@ void calculate_drop_rate(char* route_id, char* packet_size, char* maxflow_handle
       std::cout << "Maxflow integer: " << maxflow << std::endl;
       std::cout << "Totalflow integer: " << totalflow << std::endl;
       std::cout << "New drop rate: " << drop_rate << std::endl;
-      maxflow_map[route_id] = std::make_tuple(time(0), 0, drop_rate);
+      maxflow_map[route_id] = std::make_tuple(time(0), 0.0, drop_rate);
     }
   }
 }
